chore: Adjust default values for KeyboardAccelerator properties
